separate doctolib center parsers in distinct file
